% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Percent}
\usage{
percent(x = numeric())
}
\arguments{
\item{x}{Percentages.}
}
\value{
A class of object \code{percent}.
}
\description{
A lightweight class for printing proportions as percentages.
}
\examples{
library(percent)

percent(0:10)

# Convert proportions to percentages
percent(seq(0, 1, 0.1) * 100)

# You can use round() as usual
p <- percent(15.56)
round(p)
round(p, digits = 1)

signif(p, 2)
floor(p)
ceiling(p)

# We can do basic math operations as usual
10 * percent(c(0, 50, 200))
percent(10) + percent(20)

# Formatting options
format(percent(2.674), digits = 2, symbol = " (\%)")
# Prints nicely in data frames (and tibbles)
x <- seq(0, 1, 0.1)
df <- data.frame(row.names = seq_along(x))
df$perc <- percent(x)
df

}
